1リクエスト1トランザクションになるように修正

+++ conf/db/migration/default/V1__create_user_table.sql
@@ -1,5 +1,5 @@
 create table users (
-  id    bigint  primary key,
+  id    SERIAL  primary key,
   user_identifier   varchar(100),
   created_time  timestamp,
   updated_time  timestamp
